+++ typedAPI/__init__.py
@@ -4,7 +4,7 @@
 
 from .resource_path import ResourcePath
 from .headers import Headers
-from .server import Server
+from .server.main import Server
 from .response import normalise_response
