+++ typedAPI/http/headers/schema.py
@@ -1,0 +1,21 @@
+import pydantic
+
+
+
+
+class Headers:
+    def __init__(self, headers: dict):
+        self.headers = headers
+
+
+
+        
+
+    
+    
+
+
+    
+    
+
+
