+++ typedAPI/http/headers/service.py
@@ -1,0 +1,7 @@
+
+
+import starlette.requests
+
+def parse(request: starlette.requests.Request):
+    pass
+
