+++ typedAPI/http/resource_path/data.py
@@ -1,0 +1,17 @@
+# type: ignore
+
+type_map = {
+    'int': int,
+    'float': float,
+    'complex': complex,
+    'bool': bool,
+    'str': str,
+    'bytes': bytes,
+    'bytearray': bytearray,
+    'list': list,
+    'tuple': tuple,
+    'set': set,
+    'frozenset': frozenset,
+    'dict': dict,
+}
+
