+++ typedAPI/http/response/factory.py
@@ -1,0 +1,7 @@
+
+
+import starlette.responses
+
+def make_empty_reponse(status: int) -> starlette.responses.Response:
+    return
+
